Move allgather PostBox to utils for reuse
