+++ cpp/src/coll/allgather.cpp
@@ -19,100 +19,6 @@
 #include <rapidsmpf/utils/misc.hpp>
 
 namespace rapidsmpf::coll {
-namespace detail {
-
-void PostBox::insert(std::unique_ptr<Chunk> chunk) {
-    std::lock_guard lock(mutex_);
-    chunks_.emplace_back(std::move(chunk));
-}
-
-void PostBox::insert(std::vector<std::unique_ptr<Chunk>>&& chunks) {
-    std::lock_guard lock(mutex_);
-    std::ranges::for_each(chunks, [&](auto&& chunk) {
-        chunks_.emplace_back(std::move(chunk));
-    });
-}
-
-void PostBox::increment_goalpost(std::uint64_t amount) {
-    goalpost_.fetch_add(amount, std::memory_order_acq_rel);
-}
-
-bool PostBox::ready() const noexcept {
-    std::lock_guard lock(mutex_);
-    return goalpost_.load(std::memory_order_acquire) == chunks_.size();
-}
-
-std::vector<std::unique_ptr<Chunk>> PostBox::extract_ready() {
-    std::lock_guard lock(mutex_);
-    std::vector<std::unique_ptr<Chunk>> result;
-    for (auto&& chunk : chunks_) {
-        if (!chunk->is_ready()) {
-            break;
-        }
-        result.emplace_back(std::move(chunk));
-    }
-    std::erase(chunks_, nullptr);
-    goalpost_.fetch_sub(result.size(), std::memory_order_relaxed);
-    return result;
-}
-
-std::vector<std::unique_ptr<Chunk>> PostBox::extract() {
-    std::lock_guard lock(mutex_);
-    goalpost_.fetch_sub(chunks_.size(), std::memory_order_relaxed);
-    return std::move(chunks_);
-}
-
-bool PostBox::empty() const noexcept {
-    std::lock_guard lock(mutex_);
-    return chunks_.empty();
-}
-
-std::size_t PostBox::spill(BufferResource* br, std::size_t amount) {
-    std::lock_guard lock(mutex_);
-    std::vector<Chunk*> spillable_chunks;
-    std::size_t max_spillable{0};
-    std::size_t total_spilled{0};
-    for (auto&& chunk : chunks_) {
-        if (chunk->memory_type() == MemoryType::DEVICE) {
-            spillable_chunks.push_back(chunk.get());
-            max_spillable += chunk->data_size();
-        }
-    }
-    auto spill_chunk = [&](Chunk* chunk) -> std::size_t {
-        auto reservation =
-            br->reserve_or_fail(chunk->data_size(), SPILL_TARGET_MEMORY_TYPES);
-        chunk->attach_data_buffer(br->move(chunk->release_data_buffer(), reservation));
-        return chunk->data_size();
-    };
-    if (max_spillable < amount) {
-        // need to spill everything.
-        for (auto&& chunk : spillable_chunks) {
-            total_spilled += spill_chunk(chunk);
-        }
-        return total_spilled;
-    }
-    std::ranges::sort(spillable_chunks, std::less{}, [](Chunk* chunk) {
-        return chunk->data_size();
-    });
-    // Try and spill the minimum number of buffers summing to the
-    // amount we need while minimising the amount of data we need to
-    // spill.
-    while (!spillable_chunks.empty()) {
-        auto pos = std::ranges::lower_bound(
-            spillable_chunks, amount, std::less{}, [](Chunk* chunk) {
-                return chunk->data_size();
-            }
-        );
-        auto chunk = pos == spillable_chunks.end() ? spillable_chunks.back() : *pos;
-        total_spilled += spill_chunk(chunk);
-        if (total_spilled >= amount) {
-            break;
-        }
-        spillable_chunks.pop_back();
-    }
-    return total_spilled;
-}
-}  // namespace detail
 
 void AllGather::insert(std::uint64_t sequence_number, PackedData&& packed_data) {
     nlocal_insertions_.fetch_add(1, std::memory_order_relaxed);
